+++ tests/integration/test_multithread.py
@@ -106,5 +106,5 @@ class MultithreadTestCase(unittest.TestCase):
         # switch to another thread
         self.debugger.cmd('thread switch 2', 'Thread #2 brought into focus')
         self.debugger.cmd('thread current', '#2 thread id')
-        self.debugger.cmd('frame switch 3', 'switch to #3')
+        self.debugger.cmd('frame switch 2', 'switch to #2')
         self.debugger.cmd('var locals', 'sum3_jh = JoinHandle<i32> {')